requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pandas==2.2.3
matplotlib==3.10.1
seaborn==0.13.2
scikit-learn==1.6.1
nltk==3.9.1
python-dotenv==1.1.0
networkx==3.4.2
google-generativeai==0.8.4
langchain==0.3.21
praw==7.8.1